% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_tpsob.R
\name{b_tpsob}
\alias{b_tpsob}
\title{Tensor-product Sobolev space basis}
\usage{
b_tpsob(
  ...,
  p = 100,
  stdize = c("scale", "box", "symbox", "none"),
  shift = NULL,
  scale = NULL
)
}
\description{
Generates features from a tensor-product Sobolev space basis for estimating
functions in a Sobolev space with dominating mixed derivatives. Basis
functions are of the form \deqn{
    \psi_{\mathbf{j}}(\mathbf{x}) = \prod_{k=1}^d \psi_{j_k}(x_k),
} where \deqn{
     \phi_1(x) = 1, \quad\text{and}\quad
     \phi_j(x) = \sqrt{2}\cos(\pi (j-1) x).
}
}
