% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_rff.R
\name{predict.b_rff}
\alias{predict.b_rff}
\title{Evaluate random Fourier features on new data}
\usage{
\method{predict}{b_rff}(object, newx, ...)
}
\arguments{
\item{object}{The result of a call to \code{\link[=b_rff]{b_rff()}}}

\item{newx}{The x values at which evaluations are required.}
}
\value{
A new \code{b_rff} object with the Fourier features evaluated at the new
data points.
}
\description{
See the generic \code{\link[=predict]{predict()}} method for the general behavior of this function.
}
