% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{qtile}
\alias{qtile}
\title{Convert data to quantiles}
\usage{
qtile(x, ties.method = c("average", "first", "last", "random", "max", "min"))
}
\arguments{
\item{x}{A numeric vector.}

\item{ties.method}{As in \code{\link[=rank]{rank()}}.}
}
\value{
A numeric vector with values between 0 and 1.
}
\description{
Rescales \code{\link[=rank]{rank()}} by the size of the data and stores relevant information for
use in \code{\link[=predict]{predict()}}
}
