% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_conv.R
\name{b_conv}
\alias{b_conv}
\title{Random convolutional features}
\usage{
b_conv(
  x,
  p = 100,
  size = 3,
  stride = size,
  activation = c("max", "mean", "ppv"),
  kernels = NULL
)
}
\arguments{
\item{x}{A list of images, each an array or matrix of pixel values between 0}

\item{p}{The number of features to generate.}

\item{size}{The size of the convolutional kernel, e.g., 3 means a 3x3 kernel.}

\item{stride}{The stride of the convolution, i.e., how many pixels to skip
between each convolution. The default is the same as \code{size} so that the
convolution areas tile the image.}

\item{activation}{The activation function to apply to the convolved images so
that each convolution kernel produces a single feature. The default \code{"max"}
is also known as "max pooling"; "mean" is average pooling, and "ppv" is
the proportion of positive pixel values after convolution.}

\item{kernels}{A matrix of convolutional kernels, where each column is a
kernel in column-major format. Overrides \code{p} and \code{size} if provided.}
}
\value{
A matrix of random convolutional features.
}
\description{
Generates random convolutional features from a list of images, where each
image is an array or matrix of pixel values. The convolutional kernels have
entries drawn iid from a standard Normal distribution.
}
\examples{
FALSE

}
